services/image/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
diffusers>=0.33.1
transformers==4.51.3
accelerate==1.3.0
bitsandbytes==0.49.2
torch==2.7.0
Pillow>=10.0.0
